Properly serialize types that only appear at function input
When serializing graphs, we check every node for named types referenced,
so that we can register them as dependencies. We were skipping this
check for the graph inputs themselves. Since types used at input are
almost always used somewhere in the graph, we never noticed this gap
until a user reported an issue with NamedTuples.

ghstack-source-id: 50b22f2c7b1f1a39c374ab125e3981a06bf9b2f7
Pull Request resolved: https://github.com/pytorch/pytorch/pull/47775